♻️ refactor: ProductCard 타입 분리
