+++ src/types/index.ts
@@ -1,0 +1,8 @@
+interface ProductCardProps {
+  productId: number;
+  productImage: string;
+  productName: string;
+  productPrice: number;
+}
+
+export type { ProductCardProps };
